requirements.txt
gspread
google-auth
pandas
google-auth-oauthlib
google-auth-httplib2
oauth2client
google-api-python-client
streamlit
scikit-learn
matplotlib
seaborn
joblib